refactor: change login to auth and use setCookie from hono
