+++ src/controllers/api.controller.ts
@@ -1,7 +1,7 @@
 import { OpenAPIHono } from '@hono/zod-openapi';
 import { helloRouter } from './hello.controller';
 import { pushRouter } from './push.controller';
-import { loginRouter } from './login.controller';
+import { loginRouter } from './auth.controller';
 import { mediaRouter } from './media.controller';
 import { openGraphScrapeRoute } from './open-graph.controller';
